[Refactor] Move CmdLancher class to header file
