+++ TestShell/TestShell.vcxproj
@@ -127,6 +127,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="CmdLancher.cpp" />
     <ClCompile Include="cmd_launcher.cpp" />
     <ClCompile Include="main.cpp">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'"> %(AdditionalOptions)</AdditionalOptions>
@@ -144,6 +145,7 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="CmdLancher.h" />
     <ClInclude Include="test.h" />
     <ClInclude Include="TestScripts.h" />
   </ItemGroup>
